backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.41
aiosqlite==0.22.1
asyncpg==0.30.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.22
python-dotenv==1.2.2
pydantic[email]==2.12.5
stripe==14.4.0
pillow==12.1.1
httpx==0.28.1
slowapi==0.1.9
resend==2.10.0
